actualizar

+++ database/migrations/2024_03_05_220758_create_producto_almacens_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('producto_id')->nullable();
             $table->unsignedBigInteger('almacen_id')->nullable();
-            $table->integer('stock')->default(0);
+            $table->double('stock')->default(0);
             $table->timestamps();
         });
     }
